added read-host with if block for confirmation

+++ DEMO.VM.Cleanup.ps1
@@ -1,2 +1,7 @@
 # removes all resources from the DEMO resourcegroup except the VNET
-Get-AzResource -ResourceGroupName 'demo' | Where-Object {$_.Type -ne "Microsoft.Network/virtualNetworks"} | Remove-AzResource
+
+$Read = Read-Host "Do you want to remove all non-vnet resources in the Resource Group 'DEMO'?  Type 'yes' to confirm."
+
+if($Read -eq "yes") {
+    Get-AzResource -ResourceGroupName 'demo' | Where-Object {$_.Type -ne "Microsoft.Network/virtualNetworks"} | Remove-AzResource
+}
